Ports: Rename the 'fromXYZPort' to 'fromPort' since it's redundant
